Rename files to include Store. Remove unused auth0Context variables
